Rename: HocLoadingCommentList 파일위치 수정
index와 같은 레벨로 이동
