+++ webapp/src/components/ComentContainer/index.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import WithProvider from 'hoc/withProvider';
 import CommentProvider from 'contexts/Comment/Comment.Provider';
 import CreateRootCommentForm from './CommentForm/Create.Root.CommentForm';
-import HocLoadingCommentList from './Hoc/Hoc.LoadingCommentList';
+import HocLoadingCommentList from './LoadingCommentList';
 import * as S from './style';
 
 export default WithProvider({ Providers: [CommentProvider], Component: CommentContainer });
